Fix updates of items in side menu
SortedList already has logic for conditionally triggering UI-updates
depending on the contents of items. It also has built-in
de-duplication functionality (like a Set).

So this commit removes unnecessary code and fixes the identity check
in the callback. This simplifies the updates to the list by making it
only `add` if the item is in the list or `remove` if the item is not
in list.

Since `remove` is no longer called as a precursor to `add` on
updates (instead only `add` is called and the SortedList handles the
logic) there is no longer any unnecessary UI-redraws of the side-menu
when items haven't changed. This fixes #57 (reset of scroll position).